Added non-persistent graphs
Two nodes can be connected but queries and persistence of those relations are yet to be implemented. So this is more or less just setup.

The representation in memory is pretty simple right now. A relation is a dictionary where the key is a vertex and its value is a list of vertices it's connected to. It is important to note that every vertex for now, must be an integer correlating with the the __relational_id property of stored nodes. A Relations class maintains two of these graphs simultaneously, one for the in-relations and others for the out ones. They mirror each other since if 1 follows 2, then 2 must be followed by 1.

Every relation is named and stored in a hash table inside the Graph class.