Persistence via CSV files
These CSV files have a special property though, they necessarily only have three fields in the form "id1, relation, id2", effectively turning this into a very simple storage scheme for graphs known as n-triples.

It is important to know that ids are the relational ids on each node which is also that node's line number in the corresponding jarray(simple abbreviation for JSON array) file.