Graph queries
Graph queries are added but right now, they are exceedingly limiting because of the sparse amount available but there definitely exists a solid framework that can be used to create them later. The currently available queries are "v" which takes a pattern object and yields all vertices matching that pattern, and "outs" which takes a relation and yields the immediately connected vertices.