A persistent storage system for objects
The JSON objects that used to live in memory, now live in storage and are only fetched when needed. We create a folder that has a variable number of files but a constant number of lines per file, known as page size. This model lets us reason about where data is stored in a very simple way, stuff like rewrite remains O(N) and our index function continues to be O(1).

Even though this is definitely a more complex model to work with, but on the plus side, we now get to only have data we need in the memory.