A new default + patterns for "in" and "out"
The new default for query is to dfs through all graphs and yield only the unique results. It is important to note that "v" does not build on previous patterns but rather starts a new one.

New optional parameters for "in" and "out" queries are added that also function as a filter.